change project structure and more pyproject rules
- use structure penmon/penmon
- more elaborate pyproject.toml with many rules
