+++ setup.cfg
@@ -1,6 +1,6 @@
 [metadata]
 name = penmon
-version = 1.5
+version = 0.5.0
 author = Sherzod RUZMETOV
 author_email = sherzodr@gmail.com
 description = Implementation of weather station that calculates daily ETo for a reference crop using Penman-Monteith equation
@@ -17,7 +17,7 @@ classifiers =
 	Topic :: Scientific/Engineering :: Hydrology
 
 [options]
-python_requires = >=3.6
+python_requires = >=3.7
 
 [options.packages.find]
 where = src
